firefox: bugfix: hide icon when moving to internal pages
Cleanup tab handling code. Now icon is refreshed every time a tab is activated,
using an sdk/tabs event unrelated to the content script, so that it works even
on pages with no content script attached.

This solves another small bug: with the previous code, if user visited 10 pages
in the same tab, there were 10 "tab activatation" events installed, so on every
activation the icon was refreshed 10 times. This could potentially create an
overhead if a user did a huge browsing session in the same tab.